// src/app/modules/user/user.interface.ts
import { Model } from 'mongoose';

export type TFullName = {
    firstName: string;
    lastName: string;
};

export type TAddress = {
    street: string;
    city: string;
    country: string;
};

export type TOrder = {
    productName: string;
    price: number;
    quantity: number;
};

export type TUser = {
    userId: string;
    username: string;
    password: string;
    fullName: TFullName;
    age: number;
    email: string;
    isActive: boolean;
    hobbies: string[];
    address: TAddress;
    isDeleted: boolean;
    orders: TOrder[];
};

export interface UserModel extends Model<TUser> {
    isUserExists(userId: string): Promise<TUser | null>;
}